Refactor txpool types into separate files
